moved timestampParam
