+++ pragma-common/src/types.rs
@@ -1,5 +1,4 @@
-use serde::{Deserialize, Deserializer, Serialize};
-use std::ops::RangeInclusive;
+use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
 #[derive(Debug)]
@@ -74,55 +73,5 @@ impl Interval {
 
     pub fn to_seconds(&self) -> i64 {
         self.to_minutes() * 60
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum TimestampParam {
-    Single(u64),
-    Range(RangeInclusive<u64>),
-}
-
-impl From<u64> for TimestampParam {
-    fn from(ts: u64) -> Self {
-        TimestampParam::Single(ts)
-    }
-}
-
-impl<'de> Deserialize<'de> for TimestampParam {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s: String = Deserialize::deserialize(deserializer)?;
-        if let Some((start, end)) = s.split_once(',') {
-            let start = start.parse().map_err(serde::de::Error::custom)?;
-            let end = end.parse().map_err(serde::de::Error::custom)?;
-            Ok(TimestampParam::Range(start..=end))
-        } else {
-            let ts = s.parse().map_err(serde::de::Error::custom)?;
-            Ok(TimestampParam::Single(ts))
-        }
-    }
-}
-
-pub fn deserialize_option_timestamp_param<'de, D>(
-    deserializer: D,
-) -> Result<Option<TimestampParam>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: Option<String> = Option::deserialize(deserializer)?;
-    if let Some(s) = s {
-        if let Some((start, end)) = s.split_once(',') {
-            let start = start.parse().map_err(serde::de::Error::custom)?;
-            let end = end.parse().map_err(serde::de::Error::custom)?;
-            Ok(Some(TimestampParam::Range(start..=end)))
-        } else {
-            let ts = s.parse().map_err(serde::de::Error::custom)?;
-            Ok(Some(TimestampParam::Single(ts)))
-        }
-    } else {
-        Ok(None)
     }
 }
